backend/requirements.txt
Flask==3.0.3
mysql-connector-python==9.0.0
Flask-Cors==4.0.1   # biar API bisa diakses dari web frontend (misal pakai fetch JS)
